fix(android): switch to using routes instead of Nav
This fixes issues with back gesture and back buttons on Android
