+++ android/.gitignore
@@ -48,6 +48,7 @@ captures/
 .idea/assetWizardSettings.xml
 .idea/dictionaries
 .idea/libraries
+.idea/deploymentTargetSelector.xml
 # Android Studio 3 in .gitignore file.
 .idea/caches
 .idea/modules.xml
