+++ src/pages/Settings/Sync/SyncImportTicketPage.vue
