+++ src/pages/Settings/Sync/SyncNamespacesPage.vue
